+++ assets/js/room_manager.js
@@ -138,14 +138,6 @@ class RoomManager {
         $('aside .rooms-list').html(resp.side_nav_html)
       })
   }
-  new_message_scroll(user_id) {
-    let at_bottom = UccUtils.is_scroll_bottom(30);
-    if (debug) { console.log('new_message_scroll', at_bottom); }
-
-    if (at_bottom || user_id == ucxchat.user_id) {
-      UccUtils.scroll_bottom();
-    }
-  }
   // add_private(elem) {
   //   let username = elem.parent().attr('data-username')
   //   if (debug) { console.log('pvt-msg button clicked...', username) }