+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * [UCX-3770] Fix JS exception when selecting from a message popup
 * Trim a message before saving it
 * [UCX-3764] Allow user to hide the first listed room
+* [UCX-3769] Fix incoming message does not scroll window down
 
 ## 0.2.3 (2018-02-02)